task-creator.py/requirements.txt
pika==1.3.2
pydantic==2.11.7
pydantic-settings==2.10.1
requests==2.32.4
loguru==0.7.3
python-dotenv==1.1.1 